Force add ember cli rails addon
For some reason without this module the ember frontend does not build
with the following message:
`(:frontend has failed to build: - Building)`
After checking with byebug the error messages in `build_monitor.rb`
(ember-cli-rails) the messages are 3 times building and nothing else.

With `ember-cli-rails-addon` the ember frontend compiles (1st time is an
error but after receiving the build successful it works) but the page
has an error as the CSRF token initialisation is not working. By
disabling this (for now) the ember frontend loads as expected.